requirements.txt
altgraph==0.17.5
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
customtkinter==5.2.2
darkdetect==0.8.0
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
greenlet==3.1.1
idna==3.11
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
Markdown==3.10
MarkupSafe==3.0.2
narwhals==1.25.1
nest-asyncio==1.6.0
numpy==2.2.2
packaging==25.0
pandas==2.2.3
pillow==10.4.0
plotly==6.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.5
retrying==1.3.4
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
tabulate==0.9.0
tkinterweb==4.11.0
tkinterweb-tkhtml==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
Werkzeug==3.0.6
yt-dlp==2025.12.8
zipp==3.21.0
